Titano-N1/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
captcha